requirements/base.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
django-anymail==10.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django==5.0.6
djangorestframework==3.15.1
drf-spectacular-sidecar==2024.6.1
drf-spectacular[sidecar]==0.27.2
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema-specifications==2023.12.1
jsonschema==4.22.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-slugify==8.0.4
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0